+++ __tests__/edit-file.test.js
@@ -3,7 +3,7 @@ const fs = require('fs');
 
 describe('fs.readFile', () => {
   it('adds read data to array readData', done => {
-    const file = `${__dirname}/files/data/person.json`;
+    const file = `${__dirname}/../files/data/person.json`;
     const readData = [];
 
     fs.readFile(file, (err, data) => {
